+++ tests/test_routing.py
@@ -17,22 +17,6 @@ def _setup():
     config.debug = False
 
 
-@pytest.fixture(scope="function")
-def gen_room(webclient, patch_urlopen):
-    def func():
-        _rooms = set(storage.GAMES.keys())
-
-        rv = webclient.flask_test_client.post(
-            url_for("routing.route_host"), headers={"Referer": "/new/"}, data={"size": 6}, follow_redirects=True
-        )
-        assert rv.status_code == 200
-        assert len(storage.GAMES) > 0
-
-        return [i for i in storage.GAMES.keys() if i not in _rooms][0]
-
-    return func
-
-
 @pytest.fixture
 def patch_random_sample():
     _old = random.sample
